contracts: merkle: Add initial Merkle tree contract
This contract will have an interface wrapped around it and be
accessed in the main contract via library calls
