+++ .github/workflows/test.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Format and Lint
         run: |
           black --check .  # Black for Python
-          poetry run cairo-format -c $(find contracts -type f -name "*.cairo")
+          poetry run cairo-format -c $(find . -type f -name "*.cairo")
 
       - name: Compile
         run: poetry run nile compile
