+++ contracts/merkle/library.cairo
@@ -1,0 +1,1 @@
+// Implementation of a Merkle state tree using the Pedersen 2-1 hash built into StarkNet
